Split API functions and API hooks into two files
This splits up the circular dependency between AuthProvider and api.ts
